Add support for tracking three types of tasks.
Task is being used to to track user's tasks.

There may be different types of tasks that users want to store,
such as todos, deadlines and events.

Three different types of tasks are added as subclasses to Task.
These classes are Todo, Deadline and Event, which has additional
properties to track additional information about these Tasks.

Polymorphic code allows us to use a single collection to store
all the user's Tasks, and use the call the same methods, thus
simplifying the code.