Add error handling for incorrect inputs
The program has no way of handling invalid input from users, and usually
results in the program crashing. We want Duke to handle expected invalid
user input, so that the program can continue to run. Incorrect inputs by
users are invalid commands, invalid argument length, and invalid
arguments (such as giving 'done' followed by a non-integer, or an out of
range value).

Three exception classes were added, two of which extends DukeException.
InvalidArgumentException is thrown when the user's argument is invalid,
and UnknownCommandException is thrown when the user's command is
invalid. The switch statement for handling commands has been moved to
parseCommand().

This allows us to separate the main working code from the error handling
code which maintains code readability and better follow the SLAP
principles.

A getter function was also added to get the current tasks count in
TaskManager.