Divide classes into packages
All classes are not under any packages, and there is a moderate amount
of files in the same directory.

This makes the code hard to manage as files and the directory structure
does not give any indication to the classes that are being used in the
project.

A duke package is created, which contains two other packages: duke.task
and duke.exceptions, to contain the two classes, Task and DukeException,
as well as their subclasses, that are being used in the main Duke
program.

This allows the classes to be organised into packages for easier
management and the directory structure gives us information about the
classes being used in the project. By grouping related classes together,
the code is easier to navigate, maintain and extend.