Update TaskManager to use ArrayList to store Tasks
Tasks are stored in a fixed size Array, and only allows the user to
store a limted amount of tasks.

TaskManager is updated to store Tasks with an ArrayList.

By storing Tasks with an ArrayList, the user is able to store an
indefinite amount of Tasks. It also removes the need to store the
size of the Array with tasksCount, as the ArrayList provides a
method to access the current size of the list.