Add support for deleting tasks from the list
TaskManager does not have support for deleting Tasks.

A new method for deleting Tasks was added to TaskManager, which removes
the Task using ArrayList's remove method. A new command was added to
Duke, which calls the new method in TaskManager to remove the Task
at the given index.

This allows the user to remove the task, and TaskManager handles the
Tasks related functionality, while Duke handles the command related
functionality.