Update methods in Duke to non-static
Currently, methods in Duke are static, and static methods should be
avoided in an OOP application as they are considered to be procedural.

The static keyword has been removed from all methods in the Duke class.

As code has been previously refactored into the run method, and Duke
runs in an instance, we no longer need static methods and static
variables, which makes the code more testable.