Refactor creating commands into separate functions
The switch statement in Parser.parseCommand is too long, and thus the
readability of the code is reduced. This also causes the function to
be lengthy.

Let's refactor non-trivial functionality of the switch statement into
its own function that returns a Command.

This makes the code more readable and adhere to SLAP principles.

Checking of validity of indexes in arguments such as done and delete is
refactored into TaskManager from Parser as it is a functionality that
should be managed by the TaskManager instead.