Remove redundant parameters in TaskManager methods
Some functions in TaskManager methods have a boolean parameter
shouldPrintMessage that decides if the output should be printed to the
user. This was previously added as loading from file uses these methods
and thus the output should not be shown to the user. However, this is
now deprecated and should be removed from the code.

This boolean parameter has been removed from methods in TaskManager that
contain it.

The removal of deprecated code leads to better readability.