Refactor expected/min argument length as constants
Expected and minimum argument length for each command are used in the
code as magic numbers, which may confuse readers of the code.

These have been refactored as constants in their corresponding Command
subclass.

This improves readability of the code by removing magic numbers.

checkMinimumArgumentsLength and checkExactArgumentsLength has been
renamed to checkMinArgumentLength and checkExactArgumentLength to
better reflect checking a singular argument string, and to make the
variable name shorter.