Update Deadline and Event to accept date and time
Deadline and Event stores the deadline and event as strings, which is
fine for the current set of functionality but may be hard to extend the
program to emulate calendar functionality.

The String field in Deadline and Event were changed to LocalDateTime
objects, and the string containing the date and time is parsed in the
constructor of the Deadline and Event objects before they are stored.

This allows the Deadline and Event objects to store LocalDateTime
objects, which allows more functionality to be added by using the
LocalDateTime API. By parsing the LocalDateTime in the constructors,
we can prevent code duplication as Storage calls the constructor when
loading Tasks from file. This is achieved by storing the LocalDateTime
in the same string format as it was parsed from.