Add DatedTask interface for tasks with datetime
Filtering is needed to check if a Task is a Deadline or Event to extract
LocalDateTime objects from them. This makes the code hard to extend when
adding new functionality to dated Tasks, or when checking if a Task has
a LocalDateTime field.

An interface called DatedTask has been added which is implemented by
Tasks with a LocalDateTime field.

This enables easier checking of Tasks with LocalDateTime fields and
methods can be added to the interface to extract the field from objects
that implement this interface.